requirements.txt
absl-py==2.3.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.2
fsspec==2025.5.1
idna==3.10
itsdangerous==2.2.0
jax==0.6.1
jaxlib==0.6.1
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
protobuf==4.25.8
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.3
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
sounddevice==0.5.2
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.152
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
